Adds a server side check to check for an incorrectly passed price
If a user makes an API call to create a new product with an invalid
price passed, the record should not be saved in the DB with a 0
price(this is how it’s currently handled). An appropriate error
response message needs to be returned. This commit ensure that this is
taken care of with appropriate specs being added wherever applicable.
This also makes the API more consistent with the UI because currently
in the UI one cannot specify any alphabets for the product price field.